[Misc][Merv] Transfer files to root
