+++ index.html
@@ -105,7 +105,7 @@
       </div>
     </footer>
 
-    <script src="contributorList.js" type="module"></script>
-    <script src="projectList.js" type="module"></script>
+    <script src="js/ContributorList.js" type="module"></script>
+    <script src="js/ProjectList.js" type="module"></script>
   </body>
 </html>
