+++ js/constants.js
@@ -1,0 +1,106 @@
+export const github_usernames = {
+  ADORA: "Gameldon2",
+  ALEGARBES: "deandoiii",
+  ANGEL: "M4monski",
+  ARIAS: "arariadne",
+  ARMENION: "NE0FEL1S",
+  ATIENZA: "jimmm00",
+  AÑANA: "mantequilla45",
+  BAGUIO: "EarlLawrence7",
+  BASCON: "Theo12335",
+  BORNEA: "Marbench",
+  CAMUS: "Adrianne-John",
+  CASES: "LoydAMT",
+  CELESTRA: "Raven-20",
+  DENAMPO: "Erudi787",
+  LEUNG: "rhongobongo",
+  MIER: "Mier03",
+  OLPOC: "j1nsol",
+  PENASO: "jpenaso10",
+  PENIONAL: "johannamae19",
+  PEPINO: "NiJyuKyu",
+  QUIÑONES: "itzKitt",
+  RADEN: "radenabijah",
+  ROZ: "YONG0203",
+  SANTILLAN: "laurenx122",
+  SUNGAHID: "RainelynSungahid",
+  TABOADA: "phyveric24",
+  TAMPUS: "jdesolate",
+  TOROY: "MarkToroy",
+  VEGA: "Wl-lO",
+  VILLANUEVA: "kisetzuu",
+};
+
+export const projects = [
+  {
+    name: "TaskTrack",
+    contributors: [
+      github_usernames["PEPINO"],
+      github_usernames["CELESTRA"],
+      github_usernames["PENIONAL"],
+      github_usernames["QUIÑONES"],
+    ],
+    github_repo: "https://github.com/NiJyuKyu/TaskTrack",
+  },
+  {
+    name: "EDULINK",
+    contributors: [
+      github_usernames["OLPOC"],
+      github_usernames["ATIENZA"],
+      github_usernames["VEGA"],
+      github_usernames["BORNEA"],
+    ],
+    github_repo: "https://github.com/Wl-lO/edulink",
+  },
+  {
+    name: "GOURMET GAMBLE",
+    contributors: [
+      github_usernames["CASES"],
+      github_usernames["ANGEL"],
+      github_usernames["ARIAS"],
+      github_usernames["ARMENION"],
+    ],
+    github_repo: "https://github.com/LoydAMT/gourmetgamble0",
+  },
+  {
+    name: "Productivity Tracker",
+    contributors: [
+      github_usernames["ALEGARBES"],
+      github_usernames["PENASO"],
+      github_usernames["BAGUIO"],
+      github_usernames["RADEN"],
+      github_usernames["TABOADA"],
+    ],
+    github_repo: "https://github.com/jpenaso10/ProductivityTracker",
+  },
+  {
+    name: "ShopSavvy",
+    contributors: [
+      github_usernames["VILLANUEVA"],
+      github_usernames["ADORA"],
+      github_usernames["LEUNG"],
+      github_usernames["TOROY"],
+    ],
+    github_repo: "https://github.com/kisetzuu/ShopSavvy",
+  },
+  {
+    name: "TeknoySpace",
+    contributors: [
+      github_usernames["MIER"],
+      github_usernames["SANTILLAN"],
+      github_usernames["SUNGAHID"],
+      github_usernames["CAMUS"],
+    ],
+    github_repo: "https://github.com/Mier03/TeknoSpace",
+  },
+  {
+    name: "BSDoc",
+    contributors: [
+      github_usernames["DENAMPO"],
+      github_usernames["ROZ"],
+      github_usernames["BASCON"],
+      github_usernames["AÑANA"],
+    ],
+    github_repo: "https://github.com/mantequilla45/BSDOC",
+  },
+];
