+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "src",
+  "name": "msad-h2",
   "private": true,
   "version": "0.0.0",
   "type": "module",
