--- src/constants.js
@@ -1,106 +1,0 @@
-export const github_usernames = {
-  CAMUS: "Adrianne-John",
-  ARIAS: "arariadne",
-  ALEGARBES: "deandoiii",
-  BAGUIO: "EarlLawrence7",
-  DENAMPO: "Erudi787",
-  ADORA: "Gameldon2",
-  QUIÑONES: "itzKitt",
-  OLPOC: "j1nsol",
-  TAMPUS: "jdesolate",
-  ATIENZA: "jimmm00",
-  PENIONAL: "johannamae19",
-  PENASO: "jpenaso10",
-  VILLANUEVA: "kisetzuu",
-  SANTILLAN: "laurenx122",
-  CASES: "LoydAMT",
-  ANGEL: "M4monski",
-  AÑANA: "mantequilla45",
-  BORNEA: "Marbench",
-  TOROY: "MarkToroy",
-  MIER: "Mier03",
-  ARMENION: "NE0FEL1S",
-  PEPINO: "NiJyuKyu",
-  TABOADA: "phyveric24",
-  RADEN: "radenabijah",
-  SUNGAHID: "RainelynSungahid",
-  CELESTRA: "Raven-20",
-  LEUNG: "rhongobongo",
-  BASCON: "Theo12335",
-  VEGA: "Wl-lO",
-  ROZ: "YONG0203",
-};
-
-export const projects = [
-  {
-    name: "BSDoc",
-    contributors: [
-      github_usernames["DENAMPO"],
-      github_usernames["ROZ"],
-      github_usernames["BASCON"],
-      github_usernames["AÑANA"],
-    ],
-    github_repo: "https://github.com/mantequilla45/BSDOC",
-  },
-  {
-    name: "Productivity Tracker",
-    contributors: [
-      github_usernames["ALEGARBES"],
-      github_usernames["PENASO"],
-      github_usernames["BAGUIO"],
-      github_usernames["RADEN"],
-      github_usernames["TABOADA"],
-    ],
-    github_repo: "https://github.com/jpenaso10/ProductivityTracker",
-  },
-  {
-    name: "ShopSavvy",
-    contributors: [
-      github_usernames["VILLANUEVA"],
-      github_usernames["ADORA"],
-      github_usernames["LEUNG"],
-      github_usernames["TOROY"],
-    ],
-    github_repo: "https://github.com/kisetzuu/ShopSavvy",
-  },
-  {
-    name: "TaskTrack",
-    contributors: [
-      github_usernames["PEPINO"],
-      github_usernames["CELESTRA"],
-      github_usernames["PENIONAL"],
-      github_usernames["QUIÑONES"],
-    ],
-    github_repo: "https://github.com/NiJyuKyu/TaskTrack",
-  },
-  {
-    name: "TeknoySpace",
-    contributors: [
-      github_usernames["MIER"],
-      github_usernames["SANTILLAN"],
-      github_usernames["SUNGAHID"],
-      github_usernames["CAMUS"],
-    ],
-    github_repo: "https://github.com/Mier03/TeknoSpace",
-  },
-  {
-    name: "EDULINK",
-    contributors: [
-      github_usernames["OLPOC"],
-      github_usernames["ATIENZA"],
-      github_usernames["VEGA"],
-      github_usernames["BORNEA"],
-    ],
-    github_repo: "https://github.com/Wl-lO/edulink",
-  },
-  {
-    name: "GOURMET GAMBLE",
-    contributors: [
-      github_usernames["CASES"],
-      github_usernames["ANGEL"],
-      github_usernames["ARIAS"],
-      github_usernames["ARMENION"],
-    ],
-    github_repo: "https://github.com/LoydAMT/gourmetgamble0",
-  },
-];
