--- test-prs/Angel.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Blooms</title>
-</head>
-<body>
-    <hr>
-    <h1>HI, MY NAME IS KIT. MIKHA IS MY IDOL</h1>
-    <hr>
-    <br>
-    <img width="400px" height="400px"src="MIKHA.jpg" alt="MIKHA">
-
-</body>
-</html>
