--- test-prs/CAMUS.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Camus</title>
-</head>
-<body>
-    <h1 id="name">
-        ADRIANNE JOHN G. CAMUS
-    </h1>
-    <h2 class="yrsec">
-        BSCpE-2 (H2)
-    </h2>
-    <br><br>
-    <p id="para1">
-        Lorem ipsum dolor sit amet consectetur, adipisicing elit. Dolor sed modi optio consequuntur. Animi sit enim, error necessitatibus dolorem voluptatem accusantium veritatis obcaecati a, aperiam ipsam alias quaerat, quis ab.
-    </p>
-</body>
-</html>
