--- test-prs/CAMUS.txt
@@ -1,4 +1,0 @@
-Hello Everyone! Here is me, I am Adrianne John G. Camus from BSCpE2 (H2)
-
-
-    I hope everyone is doing well during this day (June 24, 2024). Keep smiling :D
