--- test-prs/Quiñones
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Short Example</title>
-</head>
-<body>
-    <h1>Welcome to My Website</h1>
-    <p>Here are a few things you can explore:</p>
-    <ul>
-        <li><a href="#about">About Me</a></li>
-        <li><a href="#services">Services</a></li>
-        <li><a href="#contact">Contact Me</a></li>
-    </ul>
-    <p>Feel free to navigate using the links above.</p>
-</body>
-</html>
-
