--- test-prs/README.md
@@ -1,2 +1,0 @@
-# msad-h2
-MSAD git/github tutorial dummy project
