--- test-prs/SUNGAHID.TXT
@@ -1,1 +1,0 @@
-Hello, I am Rainelyn Sungahid, your new Junior Engineer of the company.
