--- test-prs/SUNGAHID.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Sungahid HTML</title>
-</head>
-<body>
-    <header> I am Rainelyn Sungahid</header>
-    <h2>Lorem ipsum dolor sit amet consectetur adipisicing elit. Aut asperiores molestiae aspernatur ex aliquam, veniam ad numquam, nam dolores voluptates autem consequuntur sunt, iusto velit nisi et est in sequi.</h2>
-    <h3>MANIFESTING SINGKO!</h3>
-</body>
-</html>
-
