--- test-prs/TAMPUS.txt
@@ -1,4 +1,0 @@
-My first commit
-Quote: I am the best - Confucius
-
-Last Name: TAMPUS
