--- test-prs/VEga.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Flexter Vega</title>
-</head>
-<body>
-    <h1>Flexter Vega</h1>
-</body>
-</html>
