--- test-prs/arias.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>BINI BLOOMS</title>
-</head>
-<body>
-    <h1>NATION'S GIRL GROUP: BINI</h1>
-    <img src="https://i.pinimg.com/736x/ac/20/56/ac2056e7acdcaaf5860098d8bee56950.jpg" alt="BINI LOVELOVE" title="BINI(tawan)">
-</body>
-</html>
