--- test-prs/baguio.txt
@@ -1,11 +1,0 @@
-pull request
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Jez changes</title>
-</head>
-<body>
-</body>
-</html>
