--- test-prs/bascon.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>BASCON</title>
-</head>
-<body>
-    <hr>
-    <h1>HAIIII</h1>
-    <h3>Piplup!(Theodore)</h3>
-    <hr>
-    <img src="piplup.png" alt="piplup">
-    
-</body>
-</html>
