--- test-prs/foodinventory-Celestra.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<title>Food Inventory</title>
-<link rel="stylesheet" href="styles.css">
-</head>
-<body>
-<div class="container">
-<h2>User Registration</h2>
-<form action="register.php" method="POST">
-<label for="username">Username:</label>
-<input type="text" id="username" name="username" required><br><br>
-<label for="password">Password:</label>
-<input type="password" id="password" name="password" required><br><br>
-<button type="submit">Register</button>
-</form>
-</div>
-<script src="script.js"></script>
-</body>
-</html>
