--- test-prs/khent.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Cases</title>
-</head>
-<body>
-    <h1>WELCOME</h1>
-    <h3>Hi! I am Khent</h3>
-    
-</body>
-</html>
