--- test-prs/leung.cs
@@ -1,6 +1,0 @@
-public class product{
-    public Guid Guid {get;}
-    public string Name {get;set;}
-    public float Price {get;set;}
-    public string Url {get;set;}
-}
