--- test-prs/mier.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>MIER</title>
-</head>
-<body>
-    <h1>WELCOME</h1>
-    <h3>Hi! I am Angelina Mier</h3>
-    
-</body>
-</html>
