--- test-prs/roz.html
@@ -1,15 +1,0 @@
- <!DOCTYPE html>
- <html lang="en">
- <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>ROZ</title>
- </head>
- <body>
-    <hr>
-    <h1>HALLO</h1>
-    <h1>CUTIE AIAH - Roz</h1>
-    <hr>
-    <img src="aiah.jpg" alt="aiah">
- </body>
- </html>
