--- test-prs/santillan.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>SANTILLAN</title>
-</head>
-<body>
-    <h1>Welcome!</h1>
-    <h2>I am Mitch Lauren A. Santillan</h2>
-    <h3>Bachelor of Science in Computer Engineering</h3>
-    <p>
-        This is my first commit.</br>
-        
-    </p>
-
-</body>
-</html>
