--- test-prs/vince.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>anana</title>
-</head>
-<body>
-    <h1>WELCOME</h1>
-    <h3>Hi! I am vince</h3>
-    
-</body>
-</html>
